# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrak
  description: |
    Personal finance tracker API. All API endpoints are prefixed with `/api/v1`.

    ## Authentication
    Protected routes require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `/auth/register` or `/auth/login`.
  version: 0.1.0-alpha.0
  contact:
    name: FinTrak
  license:
    name: AGPL v3.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /statements/parse:
    post:
      tags:
        - Statements
      summary: Parse a bank statement PDF
      description: |
        Upload a bank or credit-card statement PDF for automatic extraction.
        The PDF is forwarded to the standalone statement-parser service.
      operationId: parseStatement
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to parse
                extractor:
                  type: string
                  description: Extractor name (e.g., "sbi_cc", "icici_cc")
                password:
                  type: string
                  description: PDF password if encrypted
                date_format:
                  type: string
                  description: Date format hint for the parser
              required:
                - file
      responses:
        "200":
          description: Statement parsed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseStatementResult"
        "400":
          description: Invalid file or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Password required or incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large (max 20 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Statement parser service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /statements/extractors:
    get:
      tags:
        - Statements
      summary: List available statement extractors
      description: Get the list of available parsers from the statement-parser service
      operationId: listStatementExtractors
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of extractors
          content:
            application/json:
              schema:
                type: object
                properties:
                  extractors:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        display_name:
                          type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          description: Statement parser service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============= Paperless Integration =============
...
  /paperless/import:
    post:
      tags:
        - Paperless
      summary: Import a Paperless document
      description: Download and parse a document from Paperless-ngx for import
      operationId: importPaperlessDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperlessImportRequest"
      responses:
        "200":
          description: Document parsed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseStatementResult"
        "400":
          description: Invalid request or Paperless not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Paperless or parser service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============= Dashboard =============
...
components:
  schemas:
...
    PaperlessImportRequest:
      type: object
      required:
        - documentId
      properties:
        documentId:
          type: integer
        extractor:
          type: string
        password:
          type: string
        dateFormat:
          type: string
...
    ParseStatementTransaction:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
...
    ParseStatementResult:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/ParseStatementTransaction"
        summary:
          type: object
          additionalProperties:
            type: string
        pageCount:
          type: integer
        transactionCount:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - message
            properties:
              field:
                type: string
              tag:
                type: string
              message:
                type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/register or /auth/login